04-adk-samples/adk_sample_01_authenticated_vais/agent/requirements.txt
google-cloud-discoveryengine
google-cloud-storage 
requests 
PyYAML
python-dotenv
google-cloud-aiplatform[adk,agent_engines]==1.103.0
google-auth-oauthlib
google-genai